[TGS-73] Removed notifications from BaseLibrary
