--- Source/TileGameServer.BaseLibrary/TileGameServer.BaseLibrary.Domain/MessageQueueNotifications/CreateGameSessionNotification.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TileGameServer.BaseLibrary.Domain.MessageQueueNotifications
-{
-    public class CreateGameSessionNotification
-    {
-        public Guid GameSessionId { get; set; }
-    }
-}
