--- Source/TileGameServer.BaseLibrary/TileGameServer.BaseLibrary.Domain/MessageQueueNotifications/LeaveGameSessionNotification.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TileGameServer.BaseLibrary.Domain.MessageQueueNotifications
-{
-    public class LeaveGameSessionNotification
-    {
-        public Guid PlayerId { get; set; }
-    }
-}
