+++ Source/TileGameServer.InSession/TileGameServer.InSession/Commands/Notifications/CreateGameSession/CreateGameSessionNotificationCommand.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using MediatR;
-using TileGameServer.BaseLibrary.Domain.MessageQueueNotifications;
 
 namespace TileGameServer.InSession.Commands.Notifications.CreateGameSession
 {
