+++ Source/TileGameServer/TileGameServer/Commands/Menu/Notifications/LeaveGameSession/LeaveGameSessionNotificationHandler.cs
@@ -1,7 +1,7 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using TileGameServer.BaseLibrary.Domain.MessageQueueNotifications;
 using TileGameServer.DataAccess.Repositories.Players;
 using WebApiBaseLibrary.Enums;
 using WebApiBaseLibrary.Infrastructure.MessageQueueing;
@@ -39,5 +39,10 @@ namespace TileGameServer.Commands.Menu.Notifications.LeaveGameSession
 
             return Unit.Value;
         }
+    }
+
+    public class LeaveGameSessionNotification
+    {
+        public Guid PlayerId { get; set; }
     }
 }
